+++ src/ff7/graphics.cpp
@@ -324,22 +324,25 @@ void ff7gl_field_78(struct ff7_polygon_set *polygon_set, struct ff7_game_obj *ga
 			{
 				if(ip)
 				{
-					if(zsort) ff7_externals.sub_665793(matrix_set->matrix_world, 0, ip, polygon_set, hundred_data, group_data, game_object);
+					if (zsort)
+						ff7_externals.sub_665793(matrix_set->matrix_world, 0, ip, polygon_set, hundred_data, group_data, game_object);
 					else
 					{
 						if (matrix_set)	gl_set_worldview_matrix(matrix_set->matrix_world);
+
 						if (enable_lighting)
-						{					
-							update_view_matrix(game_object);		
+						{
+							update_view_matrix(game_object);
 							if(polygon_set->light != nullptr && game_lighting != GAME_LIGHTING_ORIGINAL)
 							{
 								struct light_data lightData;
 								fill_light_data(&lightData, polygon_set);
 								gl_draw_with_lighting(ip, polygon_set->polygon_data, &lightData, polygon_set->field_4);
-							} else 	gl_draw_with_lighting(ip, polygon_set->polygon_data, nullptr, polygon_set->field_4);						
-						} else 
+							} else 	gl_draw_with_lighting(ip, polygon_set->polygon_data, nullptr, polygon_set->field_4);
+						}
+						else
 						{
-							if(polygon_set->light != nullptr && game_lighting != GAME_LIGHTING_ORIGINAL)
+							if(polygon_set->light != nullptr && game_lighting != GAME_LIGHTING_ORIGINAL && !ff7_disable_field_lighting)
 							{
 								struct light_data lightData;
 								fill_light_data(&lightData, polygon_set);